Add files via upload
In this iteration I did a few things differently, changed the data structure to be squared and less bias. Combined the regular season stats to feed an elo system, and used xgboost. This notebook contains lots of notes that could be helpful for the write up. I will work on implementing it into main